scripts/requirements.txt
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
google-analytics-admin>=0.20.0
google-analytics-data>=0.18.0
